refactor bbr main as a prep for pluggability
